docs: Fix sample code in the docs
If you try compiling the code in the quickstart guide, it will error out
because RunDefaultContainerAndStopOnCleanup does not take an option
argument.

Switching to RunContainerAndStopOnCleanup seems to do the trick although
I'm not sure at all whether this is the right function since this is my
first try at using this library.